+++ lab-scott/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "lab-scott",
+  "version": "1.0.0",
+  "description": "lab5-dsa",
+  "main": "server.js",
+  "directories": {
+    "test": "test"
+  },
+  "scripts": {
+    "test": "mocha"
+  },
+  "repository": {
+    "type": "git",
+    "url": "(\"https://github.com/ScottBrenden/lab-05-career-development.git\")"
+  },
+  "author": "scott",
+  "license": "MIT"
+}
